requirements.txt
langchain>=1.0.7
langchain-tavily>=0.2.13
langchain-anthropic>=1.1.0
python-dotenv
pydantic>=2.0.0
firecrawl-py>=4.10.0